requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0
mongoengine==0.27.0
pymongo==4.6.0
